better naming (use proxyAccount)

+++ tx/account/handler_test.go
@@ -8,13 +8,13 @@ import (
 )
 
 var (
-	c0    = sdk.Coins{sdk.Coin{Denom: "lino", Amount: int64(0)}}
-	c100  = sdk.Coins{sdk.Coin{Denom: "lino", Amount: int64(100)}}
-	c200  = sdk.Coins{sdk.Coin{Denom: "lino", Amount: int64(200)}}
-	c1600 = sdk.Coins{sdk.Coin{Denom: "lino", Amount: int64(1600)}}
-	c1800 = sdk.Coins{sdk.Coin{Denom: "lino", Amount: int64(1800)}}
-	c1900 = sdk.Coins{sdk.Coin{Denom: "lino", Amount: int64(1900)}}
-	c2000 = sdk.Coins{sdk.Coin{Denom: "lino", Amount: int64(2000)}}
+	c0    = sdk.Coins{sdk.Coin{Denom: "lino", Amount: 0}}
+	c100  = sdk.Coins{sdk.Coin{Denom: "lino", Amount: 100}}
+	c200  = sdk.Coins{sdk.Coin{Denom: "lino", Amount: 200}}
+	c1600 = sdk.Coins{sdk.Coin{Denom: "lino", Amount: 1600}}
+	c1800 = sdk.Coins{sdk.Coin{Denom: "lino", Amount: 1800}}
+	c1900 = sdk.Coins{sdk.Coin{Denom: "lino", Amount: 1900}}
+	c2000 = sdk.Coins{sdk.Coin{Denom: "lino", Amount: 2000}}
 )
 
 func TestFollow(t *testing.T) {
